Backend_demo/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
pyrebase4==4.5.0
schedule==1.2.1
